pyproject.toml
[project]
name = "ga-mcp-full"
description = "GA4 MCP server with full Admin API read/write access"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-analytics-data==0.20.0",
    "google-analytics-admin==0.27.0",
    "google-auth~=2.40",
    "google-auth-oauthlib~=1.2",
    "mcp>=1.24.0",
    "google-adk>=1.24.1",
    "httpx>=0.28.1",
]

[project.scripts]
ga-mcp-full = "ga_mcp.cli:main"

[build-system]
requires = ["setuptools>=82.0.0", "wheel"]
build-backend = "setuptools.build_meta"
